move manifest.Validate to internal
