+++ internal/manifestutil/manifestutil_test.go
@@ -4,6 +4,10 @@ import (
 	"bytes"
 	"io"
 	"io/fs"
+	"os"
+	"path"
+	"slices"
+	"strings"
 
 	. "gopkg.in/check.v1"
 
@@ -551,7 +555,7 @@ func (s *S) TestGenerateManifests(c *C) {
 		c.Assert(err, IsNil)
 		mfest, err := manifest.Read(&buffer)
 		c.Assert(err, IsNil)
-		err = manifest.Validate(mfest)
+		err = manifestutil.Validate(mfest)
 		c.Assert(err, IsNil)
 		contents := apachetestutil.DumpManifestContents(c, mfest)
 		c.Assert(contents, DeepEquals, test.expected)
@@ -573,4 +577,100 @@ func (s *S) TestGenerateNoManifests(c *C) {
 	n, err := reader.Read(bs)
 	c.Assert(err, IsNil)
 	c.Assert(n, Equals, 0)
+}
+
+var validateManifestTests = []struct {
+	summary string
+	input   string
+	mfest   *apachetestutil.ManifestContents
+	error   string
+}{{
+	summary: "Slice not found",
+	input: `
+		{"jsonwall":"1.0","schema":"1.0","count":1}
+		{"kind":"content","slice":"pkg1_manifest","path":"/manifest/manifest.wall"}
+	`,
+	error: `invalid manifest: content path "/manifest/manifest.wall" refers to missing slice pkg1_manifest`,
+}, {
+	summary: "Package not found",
+	input: `
+		{"jsonwall":"1.0","schema":"1.0","count":1}
+		{"kind":"slice","name":"pkg1_manifest"}
+	`,
+	error: `invalid manifest: slice pkg1_manifest refers to missing package "pkg1"`,
+}, {
+	summary: "Path not found in contents",
+	input: `
+		{"jsonwall":"1.0","schema":"1.0","count":1}
+		{"kind":"path","path":"/dir/","mode":"01777","slices":["pkg1_myslice"]}
+	`,
+	error: `invalid manifest: path /dir/ has no matching entry in contents`,
+}, {
+	summary: "Content and path have different slices",
+	input: `
+		{"jsonwall":"1.0","schema":"1.0","count":3}
+		{"kind":"content","slice":"pkg1_myotherslice","path":"/dir/"}
+		{"kind":"package","name":"pkg1","version":"v1","sha256":"hash1","arch":"arch1"}
+		{"kind":"path","path":"/dir/","mode":"01777","slices":["pkg1_myslice"]}
+		{"kind":"slice","name":"pkg1_myotherslice"}
+	`,
+	error: `invalid manifest: path /dir/ and content have diverging slices: \["pkg1_myslice"\] != \["pkg1_myotherslice"\]`,
+}, {
+	summary: "Content not found in paths",
+	input: `
+		{"jsonwall":"1.0","schema":"1.0","count":3}
+		{"kind":"content","slice":"pkg1_myslice","path":"/dir/"}
+		{"kind":"package","name":"pkg1","version":"v1","sha256":"hash1","arch":"arch1"}
+		{"kind":"slice","name":"pkg1_myslice"}
+	`,
+	error: `invalid manifest: content path /dir/ has no matching entry in paths`,
+}, {
+	summary: "Malformed jsonwall",
+	input: `
+		{"jsonwall":"1.0","schema":"1.0","count":1}
+		{"kind":"content", "not valid json"
+	`,
+	error: `invalid manifest: cannot read manifest: unexpected end of JSON input`,
+}}
+
+func (s *S) TestManifestValidate(c *C) {
+	for _, test := range validateManifestTests {
+		c.Logf("Summary: %s", test.summary)
+
+		// Reindent the jsonwall to remove leading tabs in each line.
+		lines := strings.Split(strings.TrimSpace(test.input), "\n")
+		trimmedLines := make([]string, 0, len(lines))
+		for _, line := range lines {
+			trimmedLines = append(trimmedLines, strings.TrimLeft(line, "\t"))
+		}
+		test.input = strings.Join(trimmedLines, "\n")
+		// Assert that the jsonwall is valid, for the test to be meaningful.
+		slices.Sort(trimmedLines)
+		orderedInput := strings.Join(trimmedLines, "\n")
+		c.Assert(test.input, DeepEquals, orderedInput, Commentf("input jsonwall lines should be ordered"))
+
+		tmpDir := c.MkDir()
+		manifestPath := path.Join(tmpDir, "manifest.wall")
+		w, err := os.OpenFile(manifestPath, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, 0644)
+		c.Assert(err, IsNil)
+		_, err = w.Write([]byte(test.input))
+		c.Assert(err, IsNil)
+		w.Close()
+
+		r, err := os.OpenFile(manifestPath, os.O_RDONLY, 0644)
+		c.Assert(err, IsNil)
+		defer r.Close()
+
+		mfest, err := manifest.Read(r)
+		c.Assert(err, IsNil)
+		err = manifestutil.Validate(mfest)
+		if test.error != "" {
+			c.Assert(err, ErrorMatches, test.error)
+			continue
+		}
+		c.Assert(err, IsNil)
+		if test.mfest != nil {
+			c.Assert(apachetestutil.DumpManifestContents(c, mfest), DeepEquals, test.mfest)
+		}
+	}
 }
